bg/lwp-619_alpa_beta: fix Scalar's getters
